backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
Pillow==10.1.0
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==1.1.0
fpdf2==2.7.6
langdetect==1.0.9
textblob==0.17.1
yake==0.4.8
slowapi==0.1.9
pydantic-settings==2.1.0
aiofiles==23.2.1
bcrypt==4.0.1
deep-translator==1.11.4
easyocr==1.7.2

# ── IndicTrans2 (AI4Bharat) — Indian language translation ──
# These are needed for the IndicTrans2 neural MT model.
# If not installed, deep-translator (Google) is used as fallback.
transformers>=4.36.0
torch>=2.1.0
sentencepiece>=0.1.99
